backend/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi
greenlet==2.0.1
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
typing-extensions==4.4.0
uvicorn
requests==2.31.0
feedparser
beautifulsoup4==4.12.2
bcrypt
passlib
PyJWT
newsdataapi
python-dotenv